requirements.txt
aiohttp>=3.9
pyyaml>=6
